+++ src/crypto/ecdsa.rs
@@ -1,4 +1,6 @@
+use ecdsa::SigningKey;
 use elliptic_curve::sec1::ToEncodedPoint;
+use p521::NistP521;
 use rand::{CryptoRng, Rng};
 use signature::hazmat::{PrehashSigner, PrehashVerifier};
 
@@ -42,6 +44,20 @@ pub fn generate_key<R: Rng + CryptoRng>(
             ))
         }
 
+        ECCCurve::P521 => {
+            let secret = p521::SecretKey::random(rng);
+            let public = secret.public_key();
+            let secret = Mpi::from_raw_slice(secret.to_bytes().as_slice());
+
+            Ok((
+                PublicParams::ECDSA(EcdsaPublicParams::P521 {
+                    key: public,
+                    p: Mpi::from_raw_slice(public.to_encoded_point(false).as_bytes()),
+                }),
+                PlainSecretParams::ECDSA(secret),
+            ))
+        }
+
         ECCCurve::Secp256k1 => {
             let secret = k256::SecretKey::random(rng);
             let public = secret.public_key();
@@ -63,10 +79,20 @@ pub fn generate_key<R: Rng + CryptoRng>(
 /// Verify an ECDSA signature.
 pub fn verify(
     p: &EcdsaPublicParams,
-    _hash: HashAlgorithm,
+    hash: HashAlgorithm,
     hashed: &[u8],
     sig: &[Mpi],
 ) -> Result<()> {
+    if let Some(field_size) = p.secret_key_length() {
+        // Error out for size mismatches that would get rejected in ecdsa::hazmat::bits2field
+        ensure!(
+            hashed.len() >= field_size / 2,
+            "Hash algorithm {:?} cannot be combined with key {:?}",
+            hash,
+            p
+        );
+    }
+
     match p {
         EcdsaPublicParams::P256 { key, .. } => {
             const FLEN: usize = 32;
@@ -111,6 +137,29 @@ pub fn verify(
 
             Ok(())
         }
+        EcdsaPublicParams::P521 { key, .. } => {
+            const FLEN: usize = 66;
+            ensure_eq!(sig.len(), 2);
+
+            let r = sig[0].as_bytes();
+            let s = sig[1].as_bytes();
+
+            ensure!(r.len() <= FLEN, "invalid R (len)");
+            ensure!(s.len() <= FLEN, "invalid S (len)");
+
+            let mut sig_bytes = [0u8; 2 * FLEN];
+
+            // add padding if the values were encoded short
+            sig_bytes[(FLEN - r.len())..FLEN].copy_from_slice(r);
+            sig_bytes[FLEN + (FLEN - s.len())..].copy_from_slice(s);
+
+            let pk = p521::ecdsa::VerifyingKey::from_affine(key.as_affine().to_owned())?;
+            let sig = p521::ecdsa::Signature::try_from(&sig_bytes[..])?;
+
+            pk.verify_prehash(hashed, &sig)?;
+
+            Ok(())
+        }
         EcdsaPublicParams::Secp256k1 { key, .. } => {
             const FLEN: usize = 32;
             ensure_eq!(sig.len(), 2);
@@ -140,9 +189,26 @@ pub fn verify(
 /// Sign using ECDSA
 pub fn sign(
     secret_key: &ECDSASecretKey,
-    _hash: HashAlgorithm,
+    hash: HashAlgorithm,
     digest: &[u8],
 ) -> Result<Vec<Vec<u8>>> {
+    if let Some(field_size) = secret_key.secret_key_length() {
+        // We require that the signing key length is matched by the hash digest length, see
+        // https://www.ietf.org/archive/id/draft-ietf-openpgp-crypto-refresh-13.html#name-algorithm-specific-fields-for-d
+
+        let field_size = match field_size {
+            66 => 64, // nist p521 is treated as though it were 512 bit-sized
+            s => s,
+        };
+
+        ensure!(
+            digest.len() >= field_size,
+            "Hash digest size ({:?}) must at least match key size ({:?})",
+            hash,
+            secret_key
+        );
+    }
+
     let (r, s) = match secret_key {
         ECDSASecretKey::P256(secret_key) => {
             let secret = p256::ecdsa::SigningKey::from(secret_key);
@@ -153,6 +219,13 @@ pub fn sign(
         ECDSASecretKey::P384(secret_key) => {
             let secret = p384::ecdsa::SigningKey::from(secret_key);
             let signature: p384::ecdsa::Signature = secret.sign_prehash(digest)?;
+            let (r, s) = signature.split_bytes();
+            (r.to_vec(), s.to_vec())
+        }
+        ECDSASecretKey::P521(secret_key) => {
+            let secret: SigningKey<NistP521> = secret_key.into();
+            let signing_key = p521::ecdsa::SigningKey::from(secret);
+            let signature: p521::ecdsa::Signature = signing_key.sign_prehash(digest)?;
             let (r, s) = signature.split_bytes();
             (r.to_vec(), s.to_vec())
         }